refactor(trampoline): extract layout/path resolution into submodule
`crates/soldr-cli/src/trampoline.rs` was 1086 LOC, past the soldr-wide
1000-LOC warn threshold. Move the self-contained "Layout / binary /
sidecar path resolution" section (Layout struct + compute_layout +
resolve_target_dir + effective_target_triple + absolutize +
find_nearest_manifest + resolve_bin_name, ~190 lines) into
`trampoline_layout.rs`. Re-exported via `pub(crate) use` so the rest
of the trampoline module continues referencing the same names.

```
before:  trampoline.rs                1086 LOC
after:   trampoline.rs                 901 LOC   ← under warn
         trampoline_layout.rs          202 LOC   ← new, under warn
```

Same pattern as the cli_args extraction landed earlier in #499. No
functional change.
